move CEMIFrame module to `cemi` package
