+++ xknx/cemi/__init__.py
@@ -1,3 +1,5 @@
 """Module for handling CEMI Frames."""
 # flake8: noqa
+from .cemi_frame import CEMIFrame
 from .cemi_handler import CEMIHandler
+from .const import CEMIFlags, CEMIMessageCode
